% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methylation_variance.R
\name{Chr_region_process}
\alias{Chr_region_process}
\title{distinguish chrXX:xxx-xxx as chrXX, xxx, xxx}
\usage{
Chr_region_process(file_tmp, method)
}
\arguments{
\item{file_tmp}{Chromosome physical segment data, chrXX:xxx-xxx format or chrXX, xxx, xxx format}
}
\value{
Processed physical fragment data, chrXX, xxx, xxx format or chrXX:xxx-xxx format
}
\description{
distinguish chrXX:xxx-xxx as chrXX, xxx, xxx
}
