% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/met_acc.R
\name{Cov_to_data}
\alias{Cov_to_data}
\title{Cov files to text files}
\usage{
Cov_to_data(cov_file, cov_file_data, chr_tmp, region, suffixname, datatmp)
}
\arguments{
\item{file_dir}{cov format data generated by bismark}

\item{region_split_file}{Processed physical fragment data, chrXX, xxx, xxx format}
}
\value{
DNA methylation or chromatin accessibility matrix
}
\description{
Cov files to text files
}
