% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/met_acc.R
\name{Methlevel_diff}
\alias{Methlevel_diff}
\title{Differential level data processing}
\usage{
Methlevel_diff(methlevel_data, group_data, suff)
}
\arguments{
\item{methlevel_data}{Chromosome-level locus data}

\item{group_data}{Group data}

\item{suff}{Group information}
}
\description{
Differential level data processing
}
