% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/met_acc_variance.R
\name{Methlevel_group_variance_analysis}
\alias{Methlevel_group_variance_analysis}
\title{Methlevel group variance analysis}
\usage{
Methlevel_group_variance_analysis(file_data, file_group_tmp, method = "t.test")
}
\arguments{
\item{file_data}{DNA methylation or chromatin accessibility level data}

\item{file_group_tmp}{Group Information}

\item{method}{Statistical methods}
}
\value{
Difference analysis results
}
\description{
Methlevel group variance analysis
}
