% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methylation.R
\name{QC_meth_unmeth}
\alias{QC_meth_unmeth}
\title{Data quality control}
\usage{
QC_meth_unmeth(file_tmp, flag = "col")
}
\arguments{
\item{file_tmp}{DNA methylation or chromatin accessibility matrix}

\item{flag}{Choose to perform quality control by column or row}
}
\value{
Data after quality control
}
\description{
Data quality control
}
