% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methylation_variance.R
\name{Meth_group_variance_analysis}
\alias{Meth_group_variance_analysis}
\title{Group analysis}
\usage{
Meth_group_variance_analysis(file_data, file_group_tmp, method = "t.test")
}
\arguments{
\item{file_data}{DNA methylation or chromatin accessibility level data}

\item{file_group_tmp}{Group Information}

\item{method}{Statistical methods}
}
\value{
Difference analysis results
}
\description{
Group analysis
}
